Crate pdsl_model

Macros

messages
Defines messages for contracts with less boilerplate code.

state
Define contract state with less boilerplate code.

Structs

CallData
The raw data with which a contract is being called.

ContractDecl
A contract declaration.

ContractInstance
An instance of a contract.

DeployHandler
A handler specific to deploying a smart contract.

EmptyContractState
An empty contract state.

ExecutionEnv
Provides a safe interface to an environment given a contract state.

MessageHandler
A read-only message handler.

MessageHandlerMut
A mutable message handler.

MessageHandlerSelector
A hash to identify a called function.

NoDeployArgs
A marker struct to tell that the deploy handler requires no arguments.

UnreachableMessageHandler
A message handler that shall never handle a message.

Enums

Error
Errors the may occure during message handling.

Traits

Contract
A simple interface to work with contracts.

ContractState
Types implementing this type can be used as contract state.

HandleCall
Types implementing this trait can handle contract calls.

Message
A message with an expected input type and output (result) type.

TestableContract
An interface that allows for simple testing of contracts.

Type Definitions

RawMessageHandler
A raw read-only message handler for the given message and state.

RawMessageHandlerMut
A raw mutable message handler for the given message and state.

Result
Results of message handling operations.